Backend/requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2020.6.20
click==8.0.4
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
